third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/document-fragment-nth.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description("DocumentFragment.querySelector matching nth-child of fragment.");

var fragment = document.createDocumentFragment();
for (var i = 0; i < 12; i++) {
    var child = document.createElement("span");
    if (i == 11)
        child.id = "pass";
    fragment.appendChild(child);
}

shouldBe('fragment.querySelector(":nth-child(12)").id', '"pass"');
</script>
